+++ pkg/x/x.go
@@ -1,0 +1,4 @@
+// `x` is a collection of experimental features being used within the GitHub CLI following common Go conventions.
+//
+// Anything contained is subject to change without notice until it is considered stable enough to be promoted.
+package x
